Remove redundant MCP E2E support code
Remove obsolete helper APIs and single-owner test abstractions while
preserving Functionality, Conformance, and LLM scenario coverage.
