+++ .github/workflows/e2e-mcp.yml
@@ -262,7 +262,6 @@ jobs:
     timeout-minutes: 60
     strategy:
       fail-fast: false
-      max-parallel: 20
       matrix:
         include:
           - id: http-read-only-query
