+++ test/e2e/mcp/src/test/java/org/apache/shardingsphere/test/e2e/mcp/functionality/PostgreSQLDatabaseGatewayE2ETest.java
@@ -17,18 +17,26 @@
 
 package org.apache.shardingsphere.test.e2e.mcp.functionality;
 
+import org.apache.shardingsphere.mcp.support.database.metadata.jdbc.RuntimeDatabaseConfiguration;
+import org.apache.shardingsphere.test.e2e.mcp.support.runtime.PostgreSQLRuntimeTestSupport;
 import org.apache.shardingsphere.test.e2e.mcp.support.runtime.RuntimeTransport;
 import org.apache.shardingsphere.test.e2e.mcp.support.transport.MCPInteractionPayloads;
 import org.apache.shardingsphere.test.e2e.mcp.support.transport.MCPPayloadAssertions;
 import org.apache.shardingsphere.test.e2e.mcp.support.transport.client.MCPInteractionClient;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.condition.EnabledIf;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.testcontainers.containers.GenericContainer;
 
 import java.io.IOException;
+import java.sql.SQLException;
 import java.sql.Types;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Stream;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -36,7 +44,49 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @EnabledIf("org.apache.shardingsphere.test.e2e.mcp.env.MCPE2ECondition#isDockerEnabled")
-class PostgreSQLDatabaseGatewayE2ETest extends AbstractPostgreSQLRuntimeE2ETest {
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class PostgreSQLDatabaseGatewayE2ETest extends AbstractTransportParameterizedE2ETest {
+    
+    private static final String LOGICAL_DATABASE_NAME = "postgres_db";
+    
+    private GenericContainer<?> container;
+    
+    @AfterAll
+    void tearDownContainer() {
+        if (null != container) {
+            container.stop();
+            container = null;
+        }
+    }
+    
+    @Override
+    protected void prepareRuntimeFixture() throws IOException {
+        if (!PostgreSQLRuntimeTestSupport.isDockerAvailable()) {
+            throw new IllegalStateException("Docker is required for the PostgreSQL-backed MCP Functionality E2E test.");
+        }
+        if (null != container) {
+            return;
+        }
+        GenericContainer<?> result = PostgreSQLRuntimeTestSupport.createContainer();
+        boolean success = false;
+        try {
+            result.start();
+            PostgreSQLRuntimeTestSupport.initializeDatabase(result);
+            container = result;
+            success = true;
+        } catch (final SQLException ex) {
+            throw new IOException("Failed to initialize PostgreSQL runtime fixture.", ex);
+        } finally {
+            if (!success) {
+                result.stop();
+            }
+        }
+    }
+    
+    @Override
+    protected Map<String, RuntimeDatabaseConfiguration> getRuntimeDatabases() {
+        return PostgreSQLRuntimeTestSupport.createRuntimeDatabases(container, LOGICAL_DATABASE_NAME);
+    }
     
     @ParameterizedTest(name = "{0}")
     @MethodSource("httpTransportCase")
@@ -103,6 +153,14 @@ class PostgreSQLDatabaseGatewayE2ETest extends AbstractPostgreSQLRuntimeE2ETest 
         Map<String, Object> actual = interactionClient.call("database_gateway_execute_query",
                 Map.of("database", LOGICAL_DATABASE_NAME, "schema", "public", "sql", "SELECT status FROM public.orders WHERE order_id = 1", "max_rows", 1));
         assertTrue(String.valueOf(actual.get("row_objects")).contains("NEW"));
+    }
+    
+    private static Map<String, Object> createExecuteUpdateArguments(final String schema, final String sql) {
+        return Map.of("database", LOGICAL_DATABASE_NAME, "schema", schema, "sql", sql, "execution_mode", "execute");
+    }
+    
+    private static Stream<Arguments> httpTransportCase() {
+        return FunctionalityTransportCases.httpTransportCase();
     }
     
     private Map<String, Object> findNested(final Map<String, Object> payload, final String collectionName, final String fieldName, final String expectedValue) {
