+++ test/e2e/mcp/src/test/java/org/apache/shardingsphere/test/e2e/mcp/llm/conversation/artifact/LLMConversationArtifactWriter.java
@@ -25,6 +25,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Collection;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -32,7 +33,10 @@ import java.util.Map;
  */
 public final class LLMConversationArtifactWriter {
     
-    private final LLME2ERuntimeEvidenceValidator runtimeEvidenceValidator = new LLME2ERuntimeEvidenceValidator();
+    private static final List<String> REQUIRED_SCORE_EVIDENCE_KEYS = List.of(
+            "runtimeMode", "dockerOwned", "provider", "serverRuntime", "serverImage", "serverImageId", "baseServerImage", "baseServerImageDigest",
+            "modelRepository", "modelReference", "servedModelId",
+            "modelQuantization", "modelRevision", "modelFileName", "modelSha256", "modelPackaging", "contextWindowTokens", "baseUrlOwnedByTest");
     
     /**
      * Write one conversation result.
@@ -45,7 +49,7 @@ public final class LLMConversationArtifactWriter {
      */
     public void write(final Path artifactDirectory, final Result conversationResult, final Map<String, Object> runtimeEvidence,
                       final Collection<String> sensitiveValues) throws IOException {
-        runtimeEvidenceValidator.validate(runtimeEvidence);
+        validateRuntimeEvidence(runtimeEvidence);
         writeContent(artifactDirectory.resolve("run-context.json"), JsonUtils.toJsonString(createRunContext(conversationResult, runtimeEvidence)), sensitiveValues);
         writeContent(artifactDirectory.resolve("system-prompt.md"), conversationResult.systemPrompt(), sensitiveValues);
         writeContent(artifactDirectory.resolve("question.txt"), conversationResult.scenario().question(), sensitiveValues);
@@ -64,6 +68,24 @@ public final class LLMConversationArtifactWriter {
                 "modelName", conversationResult.modelName(),
                 "runtime", runtimeEvidence,
                 "failureType", conversationResult.assertionReport().getFailureType());
+    }
+    
+    private void validateRuntimeEvidence(final Map<String, Object> runtimeEvidence) {
+        if (!Boolean.TRUE.equals(runtimeEvidence.get("scoreClosing"))) {
+            return;
+        }
+        for (String each : REQUIRED_SCORE_EVIDENCE_KEYS) {
+            if (isMissingEvidenceValue(runtimeEvidence.get(each))) {
+                throw new IllegalStateException(String.format("Missing score-closing LLM runtime evidence field `%s`.", each));
+            }
+        }
+        if (!Boolean.TRUE.equals(runtimeEvidence.get("dockerOwned")) || !Boolean.TRUE.equals(runtimeEvidence.get("baseUrlOwnedByTest"))) {
+            throw new IllegalStateException("Score-closing LLM runtime evidence must be Docker-owned and test-owned.");
+        }
+    }
+    
+    private boolean isMissingEvidenceValue(final Object value) {
+        return null == value || value instanceof String && ((String) value).isBlank();
     }
     
     private void writeContent(final Path file, final String content, final Collection<String> sensitiveValues) throws IOException {
