+++ test/e2e/mcp/src/test/java/org/apache/shardingsphere/test/e2e/mcp/support/distribution/DockerImageHttpRuntime.java
@@ -119,7 +119,7 @@ public final class DockerImageHttpRuntime implements AutoCloseable {
         }
     }
     
-    static List<String> createDockerCommand(final String imageName, final Path configFile, final String containerName) {
+    private static List<String> createDockerCommand(final String imageName, final Path configFile, final String containerName) {
         List<String> result = new LinkedList<>(List.of("docker", "run", "--rm", "--name", containerName,
                 "--add-host=host.docker.internal:host-gateway", "-p", "127.0.0.1::18088", "-e", "SHARDINGSPHERE_MCP_TRANSPORT=http"));
         if (null != configFile) {
@@ -185,7 +185,7 @@ public final class DockerImageHttpRuntime implements AutoCloseable {
         }
     }
     
-    static OptionalInt parsePublishedPort(final String output) {
+    private static OptionalInt parsePublishedPort(final String output) {
         for (String each : output.lines().toList()) {
             int separatorIndex = each.lastIndexOf(':');
             if (separatorIndex < 0 || separatorIndex == each.length() - 1) {
