+++ test/e2e/mcp/src/test/java/org/apache/shardingsphere/test/e2e/mcp/support/runtime/MySQLRuntimeTestSupport.java
@@ -153,7 +153,7 @@ public final class MySQLRuntimeTestSupport {
             String jdbcMetadataSchemaName = detectSchema(container);
             String resolvedPhysicalSchemaName = jdbcMetadataSchemaName.isEmpty() ? DATABASE_NAME : jdbcMetadataSchemaName;
             int totalOrders = querySingleInt(container, String.format(COUNT_ORDERS_SQL, resolvedPhysicalSchemaName));
-            LLMMySQLRuntimeFixture result = new LLMMySQLRuntimeFixture(container, logicalDatabase, totalOrders, createRuntimeDatabases(container, logicalDatabase));
+            LLMMySQLRuntimeFixture result = new LLMMySQLRuntimeFixture(container, totalOrders, createRuntimeDatabases(container, logicalDatabase));
             startupGuard.complete();
             return result;
         }
@@ -362,15 +362,14 @@ public final class MySQLRuntimeTestSupport {
     }
     
     @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-    @Getter
     public static final class LLMMySQLRuntimeFixture implements AutoCloseable {
         
         private final GenericContainer<?> container;
         
-        private final String schemaName;
-        
+        @Getter
         private final int totalOrders;
         
+        @Getter
         private final Map<String, RuntimeDatabaseConfiguration> runtimeDatabases;
         
         @Override
